# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: VHome API
  version: 0.1.0
  description: |
    VHome 第一阶段 API，覆盖唯一家庭初始化、会话、成员、存储位置、
    物料种类、库存批次、库存流水和 Dashboard。
servers:
  - url: /api/v1
    description: 当前 VHome 服务
...
paths:
...
  /members:
    get:
      tags: [Members]
      operationId: listMembers
      summary: 查询家庭成员
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/MemberRole"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MemberStatus"
      responses:
        "200":
          description: 成员列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Member"
    post:
      tags: [Members]
      operationId: createMember
      summary: 创建家庭成员
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberCreateRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: vhome_session
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
...
  responses:
...
    Forbidden:
      description: 没有权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: 当前资源状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: 字段验证失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RequestId:
      type: string
      examples: [req_01J6A9N4WQFD6T7QF3T7QF3T7Q]
    ErrorDetail:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
        reason:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        request_id:
          $ref: "#/components/schemas/RequestId"
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, page_size, total, total_pages]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          format: int64
        total_pages:
          type: integer
    PaginatedResponse:
      type: object
      required: [data, meta, request_id]
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: "#/components/schemas/PaginationMeta"
        request_id:
          $ref: "#/components/schemas/RequestId"
    IconType:
      type: string
      enum: [EMOJI, BUILTIN, UPLOAD]
    Icon:
      type: object
      additionalProperties: false
      required: [type, value]
      properties:
        type:
          $ref: "#/components/schemas/IconType"
        value:
          type: string
          minLength: 1
          maxLength: 255
    MemberRole:
      type: string
      enum: [OWNER, ADMIN, MEMBER]
    MemberStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    Member:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - email
        - role
        - status
        - created_at
        - updated_at
        - version
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/MemberRole"
        status:
          $ref: "#/components/schemas/MemberStatus"
        avatar:
          oneOf:
            - $ref: "#/components/schemas/Icon"
            - type: "null"
        last_login_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
...
    MemberResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data:
          $ref: "#/components/schemas/Member"
        request_id:
          $ref: "#/components/schemas/RequestId"
    MemberCreateRequest:
      type: object
      additionalProperties: false
      required: [name, email, initial_password, role]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        email:
          type: string
          format: email
        initial_password:
          type: string
          format: password
          minLength: 10
        role:
          $ref: "#/components/schemas/MemberRole"
        avatar:
          $ref: "#/components/schemas/Icon"
...
security:
  - sessionCookie: []